Move chart components to components/admin
