+++ front/app/components/admin/Chart/index.tsx
