+++ front/app/components/admin/NavigationTabs/tabsStyleConstants.ts
