+++ front/app/containers/Admin/dashboard/summary/charts/CumulativeAreaChart.tsx
@@ -33,7 +33,7 @@ import {
   GraphCardFigureContainer,
   GraphCardFigure,
   GraphCardFigureChange,
-} from '../..';
+} from 'components/admin/Chart';
 import { IResolution } from 'components/admin/ResolutionControl';
 
 // typings
