+++ front/app/containers/Admin/dashboard/summary/index.tsx
@@ -4,7 +4,7 @@ import { adopt } from 'react-adopt';
 import moment, { Moment } from 'moment';
 
 // components
-import { GraphsContainer, ControlBar, Column } from '../';
+import { GraphsContainer, ControlBar, Column } from 'components/admin/Chart';
 import ResolutionControl, {
   IResolution,
 } from 'components/admin/ResolutionControl';
