+++ front/app/modules/commercial/user_custom_fields/admin/components/GenderChart.tsx
@@ -24,7 +24,7 @@ import {
   GraphCard,
   GraphCardInner,
   PieChartStyleFixesDiv,
-} from 'containers/Admin/dashboard';
+} from 'components/admin/Chart';
 
 import { PieChart, Pie, Tooltip, Cell, ResponsiveContainer } from 'recharts';
